requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --no-emit-index-url requirements.in
#
asgiref==3.3.0            # via django
certifi==2020.6.20        # via httpx
django==3.1.3             # via -r requirements.in
h11==0.11.0               # via httpcore
httpcore==0.12.0          # via httpx
httpx==0.16.1             # via -r requirements.in
idna==2.10                # via rfc3986
petl==1.6.8               # via -r requirements.in
python-dateutil==2.8.1    # via -r requirements.in
pytz==2020.4              # via django
rfc3986[idna2008]==1.4.0  # via httpx
six==1.15.0               # via python-dateutil
sniffio==1.2.0            # via httpcore, httpx
sqlparse==0.4.1           # via django
